+++ Tests/Timer/DemoF8Timer.cs
@@ -9,11 +9,16 @@ namespace F8Framework.Tests
         void Start()
         {
             // 普通Timer,传入自身this，每1秒执行一次，延迟0秒后开始，执行3次(-1表示循环)
-            int timeid = FF8.Timer.AddTimer(this, 1f, 0f, 3, () => { LogF8.Log("tick"); }, () => { LogF8.Log("完成"); });
+            int timeid = FF8.Timer.AddTimer(this, 1f, 0f, 3,
+                () => { LogF8.Log("tick"); },
+                () => { LogF8.Log("完成"); },
+                ignoreTimeScale: true);
 
             // FrameTimer,传入自身this，每1帧执行一次，延迟0帧后开始，循环执行(-1表示循环)
-            timeid = FF8.Timer.AddTimerFrame(this, 1f, 0f, -1, () => { LogF8.Log("tick"); },
-                () => { LogF8.Log("完成"); });
+            timeid = FF8.Timer.AddTimerFrame(this, 1f, 0f, -1,
+                () => { LogF8.Log("tick"); },
+                () => { LogF8.Log("完成"); },
+                ignoreTimeScale: true);
 
             FF8.Timer.RemoveTimer(timeid); // 停止名为timeid的Timer